(Bug 2365)  Private Message a Community to Contact Its Moderators
Basic implementation of messaging a community to contact its admins.
In the end, this shall allow users to private message a community
account, and copy the pm to the administrators of the community.
Administrators can access community inboxes and send messages as
communities, which will attach a note as to which admin sent that
message.

The feature can be turned off just like private messages can be on a
personal account. The feature can be limited to only community members.